+++ CHANGELOG.md
@@ -13,12 +13,14 @@ Binaries to download latest version:
 
 Run in browser: [**Web**](https://healthpod.solidcommunity.au).
 
-## 0.2 LATEST UPDATES
+## 1.0 First stable release
 
-## 0.1 Initial usable release.
-
++ General cleanup [0.1.3 20250424 gjw]
 + Resync [0.1.2 20250423 gjw]
 + HOME: Concept dashboard and vaccinations [0.1.1 20250402 misc]
+
+## 0.1 Initial beta release.
+
 + Ready for beta release [0.1.0 20250326 gjw]
 + Deploy navigator [0.0.9 20250218 kev]
 + Refine and add import order lint [0.0.8 20250207 atangster]